# openapi.yaml
openapi: 3.1.0
...
info:
  title: MultiSpeed API
  version: 1.0.3
  summary: Multi-WAN speed-test scheduling, execution, routing, and statistics API
  description: |
    Same-origin API for MultiSpeed. There is deliberately no authentication.
    Keep the listener on loopback or a trusted network unless an authenticating
    reverse proxy protects it. Browser mutations are origin-validated.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: MultiSpeed issue tracker
    url: https://github.com/dude2k/MultiSpeed/issues
servers:
  - url: /api/v1
    description: Same-origin application API
security: []
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [Tasks]
      summary: List all speed-test tasks
      responses:
        "200":
          description: Tasks ordered by name.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                type: array
                maxItems: 10000
                items:
                  $ref: "#/components/schemas/Task"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createTask
      tags: [Tasks]
      summary: Create and immediately schedule a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task was persisted and its schedule was installed.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "415":
          $ref: "#/components/responses/UnsupportedContentType"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /tasks/validate:
    post:
      operationId: validateTransientTask
      tags: [Tasks]
      summary: Preflight an unsaved task and its exact network path
      description: Applies persisted and declared task defaults, validates the strict task/provider DTO and concrete interface/source address, then performs route and public-path validation. It never persists or executes the task.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "200":
          description: The unsaved task and selected network path are valid.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteValidation"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "415":
          $ref: "#/components/responses/UnsupportedContentType"
        "422":
          description: The task DTO, provider target, interface/source, route, or public path is invalid.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorEnvelope"
                  - $ref: "#/components/schemas/RouteValidation"
        "500":
          $ref: "#/components/responses/InternalError"
  /tasks/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
    get:
      operationId: getTask
      tags: [Tasks]
      summary: Get one task
      responses:
        "200":
          description: Task details.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      operationId: updateTask
      tags: [Tasks]
      summary: Replace and immediately reschedule a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "200":
          description: Updated task.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          $ref: "#/components/responses/UnsupportedContentType"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteTask
      tags: [Tasks]
      summary: Delete a task while preserving history
      description: Removes the scheduler entry and soft-deletes the task. Historical results remain available for analysis.
      responses:
        "204":
          description: Task deleted; historical results were preserved.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /tasks/{id}/run:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
    post:
      operationId: runTask
      tags: [Tasks]
      summary: Queue a manual run
      description: Limited to four requests per minute per client. Uses the same execution pipeline as scheduled runs.
      responses:
        "202":
          description: A queued result was created.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
  /tasks/{id}/validate:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
    post:
      operationId: validateTask
      tags: [Tasks]
      summary: Validate a task and its selected route
      responses:
        "200":
          description: Task and route are valid.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteValidation"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Task or selected route is invalid. The response is either the standard error envelope or a failed route snapshot.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorEnvelope"
                  - $ref: "#/components/schemas/RouteValidation"
  /tasks/{id}/duplicate:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
    post:
      operationId: duplicateTask
      tags: [Tasks]
      summary: Duplicate a task in disabled state
      responses:
        "201":
          description: Disabled copy with a new ID.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /tasks/{id}/next-runs:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
    get:
      operationId: getTaskNextRuns
      tags: [Tasks]
      summary: Calculate the next five task runs
      responses:
        "200":
          description: Next runs in UTC for the task's IANA timezone.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NextRuns"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /results:
    get:
      operationId: listResults
      tags: [Results]
      summary: List and filter results
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/TaskIDFilter"
        - $ref: "#/components/parameters/ProviderFilter"
        - $ref: "#/components/parameters/StatusFilter"
        - $ref: "#/components/parameters/InterfaceFilter"
        - $ref: "#/components/parameters/SourceIPFilter"
        - $ref: "#/components/parameters/ServerIDFilter"
        - $ref: "#/components/parameters/FromFilter"
        - $ref: "#/components/parameters/ToFilter"
        - $ref: "#/components/parameters/ResultSort"
        - $ref: "#/components/parameters/SortDirection"
      responses:
        "200":
          description: Paginated result set.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultPage"
        "400":
          $ref: "#/components/responses/InvalidFilter"
        "500":
          $ref: "#/components/responses/InternalError"
  /results/dashboard-summary:
    get:
      operationId: getDashboardResultSummary
      tags: [Results]
      summary: Get a bounded operational result snapshot
      description: |
        Returns the latest result of any status for every configured task and WAN path,
        up to 1,000 active runs, and the five most recent failures. Tasks and paths that
        have never run are represented with a null latest result. Large diagnostics are
        available only from the result-detail endpoint.
      responses:
        "200":
          description: Bounded dashboard result snapshot.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResultSummary"
        "500":
          $ref: "#/components/responses/InternalError"
  /results/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
    get:
      operationId: getResult
      tags: [Results]
      summary: Get one result with bounded diagnostics
      responses:
        "200":
          description: Result details.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteResult
      tags: [Results]
      summary: Delete one result
      responses:
        "204":
          description: Result deleted.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /results/delete-batch:
    post:
      operationId: deleteResultsBatch
      tags: [Results]
      summary: Delete a bounded set of results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [ids]
              properties:
                ids:
                  type: array
                  minItems: 1
                  maxItems: 500
                  uniqueItems: true
                  items:
                    $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Number of matching rows deleted.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [deleted]
                properties:
                  deleted:
                    type: integer
                    format: int64
                    minimum: 0
                    maximum: 500
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "415":
          $ref: "#/components/responses/UnsupportedContentType"
  /statistics:
    get:
      operationId: getStatistics
      tags: [Statistics]
      summary: Query raw or timezone-aware aggregate statistics
      parameters:
        - name: from
          in: query
          description: Inclusive start. Defaults to 30 days before the request.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Exclusive end. Defaults to the request time. Ranges are limited to ten years.
          schema:
            type: string
            format: date-time
        - name: granularity
          in: query
          schema:
            $ref: "#/components/schemas/StatisticsGranularity"
        - name: timezone
          in: query
          description: IANA reporting timezone; defaults to persisted settings.
          schema:
            type: string
            maxLength: 128
            example: Europe/Berlin
        - name: groupBy
          in: query
          schema:
            $ref: "#/components/schemas/StatisticsDimension"
        - $ref: "#/components/parameters/TaskIDListFilter"
        - $ref: "#/components/parameters/ProviderListFilter"
        - $ref: "#/components/parameters/InterfaceListFilter"
        - $ref: "#/components/parameters/SourceIPListFilter"
        - $ref: "#/components/parameters/ServerIDListFilter"
        - $ref: "#/components/parameters/RouteProfileIDListFilter"
        - $ref: "#/components/parameters/PublicIPListFilter"
      responses:
        "200":
          description: Normalized statistics report. Raw queries are limited to 90 days; the materialized report contains at most 5,000 output points and 1,000 comparison groups.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatisticsReport"
        "400":
          $ref: "#/components/responses/InvalidFilter"
        "413":
          $ref: "#/components/responses/StatisticsLimitExceeded"
        "500":
          $ref: "#/components/responses/InternalError"
  /interfaces:
    get:
      operationId: listInterfaces
      tags: [Network]
      summary: Get the most recent interface snapshot
      parameters:
        - name: includeLoopback
          in: query
          schema:
            type: boolean
            default: false
        - name: includeDown
          in: query
          schema:
            type: boolean
            default: false
        - name: includeVirtual
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Interface snapshot from the active Linux network namespace.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterfaceSnapshot"
        "400":
          $ref: "#/components/responses/InvalidFilter"
  /interfaces/refresh:
    post:
      operationId: refreshInterfaces
      tags: [Network]
      summary: Refresh interface state immediately
      responses:
        "200":
          description: Refreshed interface snapshot.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterfaceSnapshot"
        "500":
          $ref: "#/components/responses/InternalError"
  /route-profiles:
    get:
      operationId: listRouteProfiles
      tags: [Route profiles]
      summary: List persisted read-only route profiles
      responses:
        "200":
          description: Route profiles.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                type: array
                maxItems: 10000
                items:
                  $ref: "#/components/schemas/RouteProfile"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createRouteProfile
      tags: [Route profiles]
      summary: Create a route profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteProfileInput"
      responses:
        "201":
          description: Created profile.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteProfile"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "415":
          $ref: "#/components/responses/UnsupportedContentType"
        "422":
          $ref: "#/components/responses/ValidationError"
  /route-profiles/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
    get:
      operationId: getRouteProfile
      tags: [Route profiles]
      summary: Get one route profile
      responses:
        "200":
          description: Route profile.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteProfile"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateRouteProfile
      tags: [Route profiles]
      summary: Replace a route profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteProfileInput"
      responses:
        "200":
          description: Updated profile.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteProfile"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          $ref: "#/components/responses/UnsupportedContentType"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      operationId: deleteRouteProfile
      tags: [Route profiles]
      summary: Delete a route profile
      description: Soft-deletes an unreferenced profile while preserving historical result snapshots. Referenced profiles return HTTP 409.
      responses:
        "204":
          description: Profile deleted.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /route-profiles/{id}/validate:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
    post:
      operationId: validateRouteProfile
      tags: [Route profiles]
      summary: Perform read-only route and reachability validation
      responses:
        "200":
          description: Route matched the profile.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteValidation"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Validation completed with a mismatch or reachability failure.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteValidation"
...
  /providers/ookla/binary:
    get:
      operationId: getOoklaBinaryStatus
      tags: [Providers]
      summary: Report whether managed Ookla executable upload is enabled and installed
      responses:
        "200":
          description: Deployment-owned upload capability and installation state.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OoklaBinaryStatus"
    post:
      operationId: uploadOoklaBinary
      tags: [Providers]
      summary: Atomically validate and install an operator-supplied Ookla executable
      description: |
        Disabled unless APP_ALLOW_OOKLA_BINARY_UPLOAD=true and OOKLA_BINARY resolves
        beneath APP_DATA_DIR. The current Ookla EULA must already be accepted. The
        body must be one Linux amd64 ELF executable obtained separately by the
        operator. MultiSpeed checks its bounded size, architecture, executable
        version output, and only then atomically activates it. This endpoint executes
        the uploaded file for verification and must only be enabled on a trusted
        private network or behind an authenticating reverse proxy.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
              maxLength: 67108864
      responses:
        "201":
          description: The validated executable was installed in managed persistent storage.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OoklaBinaryInstallResult"
        "403":
          description: Managed executable upload is not enabled by deployment policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "413":
          description: The executable exceeds the 64 MiB upload limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "415":
          $ref: "#/components/responses/UnsupportedContentType"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /settings/ookla-eula:
    put:
      operationId: updateOoklaEulaAcceptance
      tags: [Settings, Providers]
      summary: Record or revoke explicit Ookla EULA acceptance
      description: |
        Updates MultiSpeed's persisted technical gate for the operator-supplied
        Ookla CLI. Acceptance does not install the CLI, grant a license, or
        replace any permission required by Ookla's current terms.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OoklaEULAAcceptanceRequest"
      responses:
        "200":
          description: Updated settings and persisted acceptance timestamp.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "415":
          $ref: "#/components/responses/UnsupportedContentType"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  headers:
    RequestID:
      description: Request correlation identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  schemas:
...
    TaskInput:
      description: On update, omitted defaultable fields retain their existing persisted task values.
      allOf:
        - $ref: "#/components/schemas/TaskFields"
      unevaluatedProperties: false
    Task:
      allOf:
        - $ref: "#/components/schemas/TaskFields"
        - type: object
          required: [id, cronExpression, timezone, serverSelectionMode, ipFamily, timeoutSeconds, preventOverlap, routeValidation, createdAt, updatedAt, lastScheduledAt, nextScheduledAt, networkPathValid, networkPathMessage]
          properties:
            id:
              $ref: "#/components/schemas/UUID"
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            lastScheduledAt:
              type: [string, "null"]
              format: date-time
            nextScheduledAt:
              type: [string, "null"]
              format: date-time
            networkPathValid:
              type: boolean
            networkPathMessage:
              type: string
              maxLength: 2048
      unevaluatedProperties: false
...